thumbnailer-basic/requirements.txt
Django==2.0.6
requests==2.19.1
Pillow==5.1.0
